NBA Data Analysis Python Code
Python code for the processing of the obtained data from NBA_data_acquision.py. Calculates the Hall of Fame probability for a test player using a model from scratch and the sklearn logisticRegression model. The code compares results from logistic model which was created from scratch to the sklearn model.